+++ app/src/main/java/nl/hnogames/domoticz/Interfaces/UtilityClickListener.java
@@ -29,5 +29,5 @@ public interface UtilityClickListener {
 
     void onLogClick(UtilitiesInfo utility, String range);
 
-    void onThermostatClick(int idx, int action, long newSetPoint);
+    void onThermostatClick(int idx, int action, double newSetPoint);
 }
